+++ src/python/pants/testutil/rule_runner.py
@@ -6,7 +6,22 @@ from dataclasses import dataclass
 from io import StringIO
 from pathlib import PurePath
 from tempfile import mkdtemp
-from typing import Any, Dict, Iterable, Mapping, Optional, Type, TypeVar, Union, cast
+from types import CoroutineType, GeneratorType
+from typing import (
+    Any,
+    Callable,
+    Dict,
+    Iterable,
+    Mapping,
+    Optional,
+    Sequence,
+    Type,
+    TypeVar,
+    Union,
+    cast,
+)
+
+from colors import blue, cyan, green, magenta, red, yellow
 
 from pants.base.build_root import BuildRoot
 from pants.base.specs_parser import SpecsParser
@@ -18,10 +33,11 @@ from pants.engine.fs import PathGlobs, PathGlobsAndRoot, Snapshot, Workspace
 from pants.engine.goal import Goal
 from pants.engine.internals.native import Native
 from pants.engine.internals.scheduler import SchedulerSession
-from pants.engine.internals.selectors import Params
+from pants.engine.internals.selectors import Get, Params
 from pants.engine.process import InteractiveRunner
 from pants.engine.rules import QueryRule, Rule
 from pants.engine.target import Target, WrappedTarget
+from pants.engine.unions import UnionMembership
 from pants.init.engine_initializer import EngineInitializer
 from pants.option.global_options import ExecutionOptions, GlobalOptions
 from pants.option.options_bootstrapper import OptionsBootstrapper
@@ -32,6 +48,11 @@ from pants.util.contextutil import temporary_dir
 from pants.util.dirutil import recursive_dirname, safe_file_dump, safe_mkdir, safe_open
 from pants.util.meta import frozen_after_init
 from pants.util.ordered_set import FrozenOrderedSet
+
+# -----------------------------------------------------------------------------------------------
+# `RuleRunner`
+# -----------------------------------------------------------------------------------------------
+
 
 _P = TypeVar("_P")
 
@@ -239,3 +260,162 @@ class RuleRunner:
         for each file.
         """
         return self.make_snapshot({fp: "" for fp in files})
+
+
+# -----------------------------------------------------------------------------------------------
+# `run_rule_with_mocks()`
+# -----------------------------------------------------------------------------------------------
+
+
+# TODO(#6742): Improve the type signature by using generics and type vars. `mock` should be
+#  `Callable[[SubjectType], ProductType]`.
+@dataclass(frozen=True)
+class MockGet:
+    product_type: Type
+    subject_type: Type
+    mock: Callable[[Any], Any]
+
+
+# TODO: Improve the type hints so that the return type can be inferred.
+def run_rule_with_mocks(
+    rule: Callable,
+    *,
+    rule_args: Optional[Sequence[Any]] = None,
+    mock_gets: Optional[Sequence[MockGet]] = None,
+    union_membership: Optional[UnionMembership] = None,
+):
+    """A test helper function that runs an @rule with a set of arguments and mocked Get providers.
+
+    An @rule named `my_rule` that takes one argument and makes no `Get` requests can be invoked
+    like so:
+
+    ```
+    return_value = run_rule_with_mocks(my_rule, rule_args=[arg1])
+    ```
+
+    In the case of an @rule that makes Get requests, things get more interesting: the
+    `mock_gets` argument must be provided as a sequence of `MockGet`s. Each MockGet takes the Product
+    and Subject type, along with a one-argument function that takes a subject value and returns a
+    product value.
+
+    So in the case of an @rule named `my_co_rule` that takes one argument and makes Get requests
+    for a product type `Listing` with subject type `Dir`, the invoke might look like:
+
+    ```
+    return_value = run_rule_with_mocks(
+      my_co_rule,
+      rule_args=[arg1],
+      mock_gets=[
+        MockGet(
+          product_type=Listing,
+          subject_type=Dir,
+          mock=lambda dir_subject: Listing(..),
+        ),
+      ],
+    )
+    ```
+
+    If any of the @rule's Get requests involve union members, you should pass a `UnionMembership`
+    mapping the union base to any union members you'd like to test. For example, if your rule has
+    `await Get(TestResult, TargetAdaptor, target_adaptor)`, you may pass
+    `UnionMembership({TargetAdaptor: PythonTestsTargetAdaptor})` to this function.
+
+    :returns: The return value of the completed @rule.
+    """
+
+    task_rule = getattr(rule, "rule", None)
+    if task_rule is None:
+        raise TypeError(f"Expected to receive a decorated `@rule`; got: {rule}")
+
+    if rule_args is not None and len(rule_args) != len(task_rule.input_selectors):
+        raise ValueError(
+            f"Rule expected to receive arguments of the form: {task_rule.input_selectors}; got: {rule_args}"
+        )
+
+    if mock_gets is not None and len(mock_gets) != len(task_rule.input_gets):
+        raise ValueError(
+            f"Rule expected to receive Get providers for {task_rule.input_gets}; got: {mock_gets}"
+        )
+
+    res = rule(*(rule_args or ()))
+    if not isinstance(res, (CoroutineType, GeneratorType)):
+        return res
+
+    def get(product, subject):
+        provider = next(
+            (
+                mock_get.mock
+                for mock_get in mock_gets
+                if mock_get.product_type == product
+                and (
+                    mock_get.subject_type == type(subject)
+                    or (
+                        union_membership
+                        and union_membership.is_member(mock_get.subject_type, subject)
+                    )
+                )
+            ),
+            None,
+        )
+        if provider is None:
+            raise AssertionError(
+                f"Rule requested: Get{(product, type(subject), subject)}, which cannot be satisfied."
+            )
+        return provider(subject)
+
+    rule_coroutine = res
+    rule_input = None
+    while True:
+        try:
+            res = rule_coroutine.send(rule_input)
+            if isinstance(res, Get):
+                rule_input = get(res.product_type, res.subject)
+            elif type(res) in (tuple, list):
+                rule_input = [get(g.product_type, g.subject) for g in res]
+            else:
+                return res
+        except StopIteration as e:
+            if e.args:
+                return e.value
+
+
+class MockConsole:
+    """An implementation of pants.engine.console.Console which captures output."""
+
+    def __init__(self, use_colors=True):
+        self.stdout = StringIO()
+        self.stderr = StringIO()
+        self.use_colors = use_colors
+
+    def write_stdout(self, payload):
+        self.stdout.write(payload)
+
+    def write_stderr(self, payload):
+        self.stderr.write(payload)
+
+    def print_stdout(self, payload):
+        print(payload, file=self.stdout)
+
+    def print_stderr(self, payload):
+        print(payload, file=self.stderr)
+
+    def _safe_color(self, text: str, color: Callable[[str], str]) -> str:
+        return color(text) if self.use_colors else text
+
+    def blue(self, text: str) -> str:
+        return self._safe_color(text, blue)
+
+    def cyan(self, text: str) -> str:
+        return self._safe_color(text, cyan)
+
+    def green(self, text: str) -> str:
+        return self._safe_color(text, green)
+
+    def magenta(self, text: str) -> str:
+        return self._safe_color(text, magenta)
+
+    def red(self, text: str) -> str:
+        return self._safe_color(text, red)
+
+    def yellow(self, text: str) -> str:
+        return self._safe_color(text, yellow)
